Split up example into more modules, fix clippy and tests, make extern functions not public
