+++ example/src/lib.rs
@@ -1,2 +1,4 @@
+pub mod data_provider;
+pub mod decimal;
 pub mod fixed_decimal;
-pub mod other;
+pub mod locale;
